Fix GoatCounter analytics
