+++ frontend/index.html
@@ -9,5 +9,11 @@
   <body>
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
+        <!-- GoatCounter -->
+    <script 
+      data-goatcounter="https://partakers-analytics.goatcounter.com/count"
+      async 
+      src="//gc.zgo.at/count.js">
+    </script>
   </body>
 </html>
